refactor(appointment): 페이지 내부 컴포넌트를 추상화
- 약속 결과 페이지 컴포넌트 내부의 컴포넌트 레벨을 추상화했다.
- 페이지 내부 컴포넌트의 추상화 레벨을 맞춤으로써, 페이지 내부의 레이아웃을 한번에 파악할 수 있다.
